+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Introduced csv2db icon
 - Check for mandatory command line arguments (ER #14)
+- Exit program immediately if no files were found (ER #17)
 
 ### Fixed
 - Pass correct exit return codes (bug #18)